bpf: tests: condense LB no-backend test
Run the IPv4 and IPv6 test from a single file.
